+++ programs/cp-amm/src/lib.rs
@@ -321,7 +321,7 @@ pub mod cp_amm {
         ctx: Context<SplitPositionCtx>,
         params: SplitPositionParameters,
     ) -> Result<()> {
-        instructions::handle_split_position2(ctx, params.get_split_position_parameters2()?)
+        instructions::handle_split_position2(ctx, params.get_split_position_parameters()?)
     }
 
     pub fn split_position2(ctx: Context<SplitPositionCtx>, numerator: u32) -> Result<()> {